Add files via upload
I created a Java FX-based Quiz Application. The goal of our project is to design an application that allows you to take a quiz several times and view the results. This idea was motivated by the university's blackboard quiz feature. This application functions similarly to our University quiz on Blackboard, allowing you to take a quiz and then view the results.